+++ src/wai/common/iterate/__init__.py
@@ -1,4 +1,5 @@
 """
 Utility functions for working with iterables and iterators.
 """
+from ._ConstantIterator import ConstantIterator
 from ._functions import safe_iter, flatten, invert_indices, extract_by_index, is_iterable, first, random